+++ backend/services/db-importer-worker/src/repositories/blob/BlobStorageGCSRepository.ts
@@ -1,4 +1,4 @@
-import { assert } from "console";
+import assert from "assert";
 import { IBlobStorageRepository } from "./IBlobStorageRepository";
 import { GetSignedUrlConfig, Storage } from '@google-cloud/storage'
 import { Logger } from "pino";
